requirements.txt
langchain==0.1.20
langchain-community==0.0.38

langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
streamlit
